Core: fix another compiler warning in libusb_setlocale()
* VS2012 (64 bit) produces the following in strerror.c(156):
  warning C4267: '=' : conversion from 'size_t' to 'int', possible loss of data

+++ libusb/strerror.c
@@ -29,7 +29,7 @@
 #define strncasecmp _strnicmp
 #endif
 
-static int usbi_locale = 0;
+static size_t usbi_locale = 0;
 
 /** \ingroup misc
  * How to add a new \ref libusb_strerror() translation:
